requirements.txt
gradio
openai-whisper
gtts
sentence-transformers
faiss-cpu
pandas
